refactor: userQuota fetching
